ci: move test to markdown_data
